Single, or multiple whitelist/blacklists allowed python-diamond/Diamond#230
I adjusted collector.py code to handle multiple whitelists or
blacklists, and used the HadoopCollector to test the changes.

The collector.process_config now ensures that
collector.config['metrics_whitelist'] (and blacklist) is either a list
of regex objects, or an empty list.

collector.metric_allowed is based on work from a previous commit and is
now generalized enough use either with path or name filtering. It now
checks white/black lists in a loop. The collector.publish uses that same
method now to avoid code duplication.

For testing I first confirmed that I broke tests by making
'metrics_blacklist' in testhadoop.py a tuple (because the code expected
a string). I then made my changes and confirmed the test was still
broken because the newly blacklisted path was not in the list of
expected metrics. I then edited expected.pkl to remove that metric, and
confirmed tests passed once again.